refactor(agent): extract CLI enum types module (#214)
